[wip] Fixes to datasets' method of working with the utils consumption
