+++ src/beacon/omop/datasets.py
@@ -1,7 +1,7 @@
 from typing import Optional
 from ...beacon.omop.filters import apply_filters
 from ...beacon.omop.schemas import DefaultSchemas
-from utils import query_id, get_count, get_documents, get_cross_query
+from ...beacon.omop.utils import query_id, get_count, get_documents, get_cross_query
 from ...beacon.request.model import RequestParams
 from ...beacon.omop import engine # client
 
@@ -10,13 +10,13 @@ import logging
 LOG = logging.getLogger(__name__)
 
 
-def get_datasets(entry_id: Optional[str], qparams: RequestParams):
+async def get_datasets(entry_id: Optional[str], qparams: RequestParams):
     collection = 'datasets'
     query = apply_filters({}, qparams.query.filters, collection)
     schema = DefaultSchemas.DATASETS
-    count = get_count(query)
+    count = get_count('beacon.datasets', query)
     docs = get_documents(
-        client.beacon.datasets,
+        'beacon.datasets',
         query,
         qparams.query.pagination.skip,
         qparams.query.pagination.limit
